push notification solved google sign in added custom posts made

+++ lib/features/workout/ui/view_workout_data.dart
@@ -45,9 +45,8 @@ class _ViewWorkoutDetailsState extends State<ViewWorkoutDetails> {
               padding: const EdgeInsets.all(8),
               itemCount: state.workoutHistory.length,
               itemBuilder: (BuildContext context, int index) {
-                WorkoutModel workoutModel = state.workoutHistory[index];
-
-                final date = DateTime.parse(workoutModel.createdAt!);
+                final workoutEntry = state.workoutHistory[index];
+                final date = DateTime.parse(workoutEntry['created_at']);
                 final formattedDate = DateFormat.yMd().add_Hm().format(date);
                 final today = DateTime.now();
                 final isToday = today.year == date.year &&
@@ -58,15 +57,77 @@ class _ViewWorkoutDetailsState extends State<ViewWorkoutDetails> {
                   elevation: 4,
                   child: ListTile(
                     title: Text(formattedDate),
-                    subtitle: Text(
-                      "Sets: ${workoutModel.sets} Reps: ${workoutModel.reps} Weight: ${workoutModel.weight}",
+                    // Inside the ListTile widget
+                    subtitle: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          "Sets: ${workoutEntry['sets']}",
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        Text(
+                          "Reps: ${workoutEntry['reps']}",
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        Text(
+                          "Weight: ${workoutEntry['weight']}",
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        Text(
+                          "Volume: ${workoutEntry['volume']}",
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        // Text(
+                        //   "Target Body Part: ${getTargetBodyParts(workoutEntry['exercise']['target_body_part'])}",
+                        //   style: TextStyle(fontWeight: FontWeight.bold),
+                        // ),
+                        Text(
+                          "Exercise Type: ${workoutEntry['exercise']['type']['name']}",
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        Text(
+                          "Exercise Name: ${workoutEntry['exercise']['exercise_name']}",
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        Text(
+                          "Calories Burned per Hour: ${workoutEntry['exercise']['calories_burned_per_hour']}",
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        // Add more fields as needed
+                      ],
                     ),
+
                     trailing: GestureDetector(
                       onTap: () {
                         if (isToday) {
-                          workoutBloc.add(
-                            DeleteWorkoutEvent(
-                                workoutID: workoutModel.workoutID!),
+                          showDialog(
+                            context: context,
+                            builder: (BuildContext context) {
+                              return AlertDialog(
+                                title: Text("Delete Workout?"),
+                                content: Text(
+                                    "Are you sure you want to delete this workout entry?"),
+                                actions: [
+                                  TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: Text("Cancel"),
+                                  ),
+                                  TextButton(
+                                    onPressed: () {
+                                      workoutBloc.add(
+                                        DeleteWorkoutEvent(
+                                            workoutID:
+                                                workoutEntry['workout_id']),
+                                      );
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: Text("Delete"),
+                                  ),
+                                ],
+                              );
+                            },
                           );
                         } else {
                           ScaffoldMessenger.of(context).showSnackBar(
